add ConvertDataUtils to testing
